requirements.txt
fastapi
uvicorn[standard]
gspread
google-auth-oauthlib
pandas
python-dotenv
